+++ server/src/main/java/com/inventage/portal/gateway/proxy/middleware/cors/CorsMiddleware.java
@@ -2,15 +2,20 @@ package com.inventage.portal.gateway.proxy.middleware.cors;
 
 import com.inventage.portal.gateway.proxy.middleware.TraceMiddleware;
 import io.opentelemetry.api.trace.Span;
+import io.vertx.core.http.HttpMethod;
 import io.vertx.ext.web.RoutingContext;
 import io.vertx.ext.web.handler.CorsHandler;
+import java.util.List;
+import java.util.Set;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 /**
  * Middleware for adding the CorsHandler of Vert.x.
  * see https://vertx.io/docs/vertx-web/java/#_cors_handling
+ * see https://developer.mozilla.org/en-US/docs/Web/HTTP/CORS
  * see https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Access-Control-Allow-Origin
+ * see https://fetch.spec.whatwg.org/#http-cors-protocol
  */
 public class CorsMiddleware extends TraceMiddleware {
 
@@ -19,9 +24,41 @@ public class CorsMiddleware extends TraceMiddleware {
     private final String name;
     private final CorsHandler corsHandler;
 
-    public CorsMiddleware(String name, String allowedOrigin) {
+    public CorsMiddleware(
+        String name,
+        List<String> allowedOrigin,
+        List<String> allowedOriginPattern,
+        Set<HttpMethod> allowedMethods,
+        Set<String> allowedHeaders,
+        Set<String> exposedHeaders,
+        int maxAgeSeconds,
+        boolean allowCredentials,
+        boolean allowPrivateNetwork
+    ) {
         this.name = name;
-        this.corsHandler = CorsHandler.create().addOrigin(allowedOrigin);
+        this.corsHandler = CorsHandler.create();
+
+        if (allowedOrigin != null && !allowedOrigin.isEmpty()) {
+            corsHandler.addOrigins(allowedOrigin);
+        }
+        if (allowedOriginPattern != null && !allowedOriginPattern.isEmpty()) {
+            corsHandler.addRelativeOrigins(allowedOriginPattern);
+        }
+        if (allowedMethods != null && !allowedMethods.isEmpty()) {
+            corsHandler.allowedMethods(allowedMethods);
+        }
+        if (allowedHeaders != null && !allowedHeaders.isEmpty()) {
+            corsHandler.allowedHeaders(allowedHeaders);
+        }
+        if (exposedHeaders != null && !exposedHeaders.isEmpty()) {
+            corsHandler.exposedHeaders(exposedHeaders);
+        }
+
+        if (maxAgeSeconds >= 0) {
+            corsHandler.maxAgeSeconds(maxAgeSeconds);
+        }
+        corsHandler.allowCredentials(allowCredentials);
+        corsHandler.allowPrivateNetwork(allowPrivateNetwork);
     }
 
     @Override