+++ server/src/main/java/com/inventage/portal/gateway/proxy/middleware/cors/CorsMiddlewareFactory.java
@@ -5,15 +5,24 @@ import com.inventage.portal.gateway.proxy.middleware.Middleware;
 import com.inventage.portal.gateway.proxy.middleware.MiddlewareFactory;
 import io.vertx.core.Future;
 import io.vertx.core.Vertx;
+import io.vertx.core.http.HttpMethod;
+import io.vertx.core.json.JsonArray;
 import io.vertx.core.json.JsonObject;
 import io.vertx.ext.web.Router;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Set;
+import java.util.stream.Collectors;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-/**
- */
 public class CorsMiddlewareFactory implements MiddlewareFactory {
 
+    public static final int DEFAULT_MAX_AGE_SECONDS = -1;
+    public static final boolean DEFAULT_ALLOW_CREDENTIALS = false;
+    public static final boolean DEFAULT_ALLOW_PRIVATE_NETWORKS = false;
+
+    private static final String ORIGIN_LOCALHOST = "http://localhost";
     private static final Logger LOGGER = LoggerFactory.getLogger(CorsMiddlewareFactory.class);
 
     @Override
@@ -23,7 +32,76 @@ public class CorsMiddlewareFactory implements MiddlewareFactory {
 
     @Override
     public Future<Middleware> create(Vertx vertx, String name, Router router, JsonObject middlewareConfig) {
+
+        final JsonArray allowedOrigin = middlewareConfig.getJsonArray(DynamicConfiguration.MIDDLEWARE_CORS_ALLOWED_ORIGINS);
+        final JsonArray allowedOriginPatterns = middlewareConfig.getJsonArray(DynamicConfiguration.MIDDLEWARE_CORS_ALLOWED_ORIGIN_PATTERNS);
+
+        final JsonArray allowedMethods = middlewareConfig.getJsonArray(DynamicConfiguration.MIDDLEWARE_CORS_ALLOWED_METHODS);
+        final JsonArray allowedHeaders = middlewareConfig.getJsonArray(DynamicConfiguration.MIDDLEWARE_CORS_ALLOWED_HEADERS);
+        final JsonArray exposedHeaders = middlewareConfig.getJsonArray(DynamicConfiguration.MIDDLEWARE_CORS_EXPOSED_HEADERS);
+
+        final int maxAgeSeconds = middlewareConfig.getInteger(DynamicConfiguration.MIDDLEWARE_CORS_MAX_AGE_SECONDS, DEFAULT_MAX_AGE_SECONDS);
+        final boolean allowCredentials = middlewareConfig.getBoolean(DynamicConfiguration.MIDDLEWARE_CORS_ALLOW_CREDENTIALS, DEFAULT_ALLOW_CREDENTIALS);
+        final boolean allowPrivateNetwork = middlewareConfig.getBoolean(DynamicConfiguration.MIDDLEWARE_CORS_ALLOW_PRIVATE_NETWORK, DEFAULT_ALLOW_PRIVATE_NETWORKS);
+
         LOGGER.info("Created '{}' middleware successfully", DynamicConfiguration.MIDDLEWARE_CORS);
-        return Future.succeededFuture(new CorsMiddleware(name, "http://localhost"));
+        return Future.succeededFuture(new CorsMiddleware(
+            name,
+            allowedOrigins(allowedOrigin),
+            allowedOriginPatterns(allowedOriginPatterns),
+            allowedMethods(allowedMethods),
+            toSet(allowedHeaders),
+            toSet(exposedHeaders),
+            maxAgeSeconds,
+            allowCredentials,
+            allowPrivateNetwork));
     }
+
+    /**
+     * Allowed origins always contains the 'http;//localhost' origin
+     * 
+     * @param originsJSON
+     * @return
+     */
+    @SuppressWarnings("unchecked")
+    private List<String> allowedOrigins(JsonArray originsJSON) {
+        if (originsJSON == null || originsJSON.isEmpty()) {
+            return List.of(ORIGIN_LOCALHOST);
+        }
+
+        final List<String> origins = originsJSON.getList();
+        if (!origins.contains(ORIGIN_LOCALHOST)) {
+            origins.add(ORIGIN_LOCALHOST);
+        }
+
+        return origins;
+    }
+
+    @SuppressWarnings("unchecked")
+    private List<String> allowedOriginPatterns(JsonArray originPatterns) {
+        if (originPatterns == null || originPatterns.isEmpty()) {
+            return List.of();
+        }
+        return originPatterns.getList();
+    }
+
+    @SuppressWarnings("unchecked")
+    private Set<HttpMethod> allowedMethods(JsonArray methods) {
+        if (methods == null || methods.isEmpty()) {
+            return Set.of();
+        }
+
+        return (Set<HttpMethod>) methods.getList().stream()
+            .map(m -> HttpMethod.valueOf((String) m))
+            .collect(Collectors.<HttpMethod>toSet());
+    }
+
+    @SuppressWarnings("unchecked")
+    private Set<String> toSet(JsonArray list) {
+        if (list == null || list.isEmpty()) {
+            return Set.of();
+        }
+        return new HashSet<String>(list.getList());
+    }
+
 }